Support single-item array in page `Contents` key
The `Contents` key can either be a direct reference to an object (stream) or an array. This PR supports the simple case, in which the array contains a single item.

For full support, the streams within the array would need to be concatinated, however, this seems to be a rather special case. This should fix most of the PDFs that failed previously.

Fixes #15
Fixes #24